requirements.txt
numpy
matplotlib
scikit-image
scikit-learn
pickles
torchvision
torch-geometric
transformers
opencv-python
#pip 

#git clone https://github.com/Cadene/skip-thoughts.torch.git
#cd skip-thoughts.torch/pytorch
#python setup.py install
#https://github.com/Cadene/skip-thoughts.torch/tree/master/pytorch